+++ novelsave/core/services/cloud/filehost/__init__.py
@@ -1,0 +1,1 @@
+from .base_cloud_file_host import BaseCloudFileHost