+++ novelsave/services/cloud/filehost/__init__.py
@@ -1,0 +1,2 @@
+from .gofiles import GoFilesHost
+from .anonfiles import AnonFilesHost